refactor: create artist layout and modify outlet
