+++ src/layouts/ArtistLayout.tsx
@@ -1,0 +1,13 @@
+import { Outlet } from "react-router-dom";
+import { ArtistSidebar } from "@pages";
+
+function ArtistLayout() {
+  return (
+    <main className="">
+      <ArtistSidebar />
+      <Outlet /> 
+    </main>
+  )
+}
+
+export default ArtistLayout
