+++ src/layouts/index.ts
@@ -1,5 +1,7 @@
-export { default as UserAccountLayout } from './UserAccountLayout';
-export { default as AdminLayout } from './AdminLayout';
+export { default as UserAccountLayout } from "./UserAccountLayout";
+export { default as AdminLayout } from "./AdminLayout";
+export { default as ArtistLayout } from "./ArtistLayout";
 
-export * from './UserAccountLayout';
-export * from './AdminLayout';
+export * from "./UserAccountLayout";
+export * from "./AdminLayout";
+export * from "./ArtistLayout";
